requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.8.1
regex==2024.5.15
dateparser==1.2.0
streamlit==1.36.0
beautifulsoup4==4.12.3
google-api-python-client==2.137.0    # only if using Gmail API
google-auth-httplib2==0.2.0          # only if using Gmail API
google-auth-oauthlib==1.2.0          # only if using Gmail API
